requirements.txt
torch
numpy
ultralytics
dill
faiss-cpu
opencv-python